Add files via upload
Utilizes a text file to take in inputs. Trace is generated by giving a range of elements to generate, for a given amount of processes. And then various options are used to decide how they're interleaved into one trace. Can still add more options but for now only sequential and random interleaving are supported.